Refactor API docs to document things more easily
