--- docs/source/API.rst
@@ -1,57 +1,0 @@
-pyresample API
-==============
-
-pyresample.geometry
--------------------
-.. automodule:: pyresample.geometry
-    :members:
-
-pyresample.image
-----------------
-.. automodule:: pyresample.image
-    :members:
-
-pyresample.grid
----------------
-.. automodule:: pyresample.grid
-    :members:
-
-pyresample.kd_tree
-------------------
-.. automodule:: pyresample.kd_tree
-    :members:
-
-pyresample.bilinear
--------------------
-.. automodule:: pyresample.bilinear
-    :members:
-
-pyresample.utils
-----------------
-.. automodule:: pyresample.utils
-    :members:
-
-pyresample.area_config
-----------------------
-.. automodule:: pyresample.area_config
-    :members:
-
-pyresample.data_reduce
-----------------------
-.. automodule:: pyresample.data_reduce
-    :members:
-
-pyresample.plot
----------------
-.. automodule:: pyresample.plot
-    :members:
-
-pyresample.ewa
---------------
-.. automodule:: pyresample.ewa
-    :members:
-
-pyresample.bucket
------------------
-.. automodule:: pyresample.bucket
-    :members:
