+++ docs/source/api/pyresample.bilinear.rst
@@ -1,0 +1,22 @@
+pyresample.bilinear package
+===========================
+
+Submodules
+----------
+
+pyresample.bilinear.xarr module
+-------------------------------
+
+.. automodule:: pyresample.bilinear.xarr
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: pyresample.bilinear
+   :members:
+   :undoc-members:
+   :show-inheritance:
