+++ docs/source/api/pyresample.bucket.rst
@@ -1,0 +1,10 @@
+pyresample.bucket package
+=========================
+
+Module contents
+---------------
+
+.. automodule:: pyresample.bucket
+   :members:
+   :undoc-members:
+   :show-inheritance:
