+++ docs/source/api/pyresample.ewa.rst
@@ -1,0 +1,10 @@
+pyresample.ewa package
+======================
+
+Module contents
+---------------
+
+.. automodule:: pyresample.ewa
+   :members:
+   :undoc-members:
+   :show-inheritance:
