+++ docs/source/api/pyresample.gradient.rst
@@ -1,0 +1,10 @@
+pyresample.gradient package
+===========================
+
+Module contents
+---------------
+
+.. automodule:: pyresample.gradient
+   :members:
+   :undoc-members:
+   :show-inheritance:
