+++ docs/source/api/pyresample.rst
@@ -1,0 +1,130 @@
+pyresample package
+==================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   pyresample.bilinear
+   pyresample.bucket
+   pyresample.ewa
+   pyresample.gradient
+   pyresample.utils
+
+Submodules
+----------
+
+pyresample.area\_config module
+------------------------------
+
+.. automodule:: pyresample.area_config
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.boundary module
+--------------------------
+
+.. automodule:: pyresample.boundary
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.data\_reduce module
+------------------------------
+
+.. automodule:: pyresample.data_reduce
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.geo\_filter module
+-----------------------------
+
+.. automodule:: pyresample.geo_filter
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.geometry module
+--------------------------
+
+.. automodule:: pyresample.geometry
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.grid module
+----------------------
+
+.. automodule:: pyresample.grid
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.image module
+-----------------------
+
+.. automodule:: pyresample.image
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.kd\_tree module
+--------------------------
+
+.. automodule:: pyresample.kd_tree
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.plot module
+----------------------
+
+.. automodule:: pyresample.plot
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.resampler module
+---------------------------
+
+.. automodule:: pyresample.resampler
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.spherical module
+---------------------------
+
+.. automodule:: pyresample.spherical
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.spherical\_geometry module
+-------------------------------------
+
+.. automodule:: pyresample.spherical_geometry
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.version module
+-------------------------
+
+.. automodule:: pyresample.version
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: pyresample
+   :members:
+   :undoc-members:
+   :show-inheritance:
