+++ docs/source/api/pyresample.utils.rst
@@ -1,0 +1,38 @@
+pyresample.utils package
+========================
+
+Submodules
+----------
+
+pyresample.utils.cartopy module
+-------------------------------
+
+.. automodule:: pyresample.utils.cartopy
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.utils.proj4 module
+-----------------------------
+
+.. automodule:: pyresample.utils.proj4
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyresample.utils.rasterio module
+--------------------------------
+
+.. automodule:: pyresample.utils.rasterio
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: pyresample.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
