+++ pyresample/utils/__init__.py
@@ -24,8 +24,8 @@ import numpy as np
 import pyproj
 import warnings
 
-from ._proj4 import (proj4_dict_to_str, proj4_str_to_dict, convert_proj_floats, proj4_radius_parameters)  # noqa
-from ._rasterio import get_area_def_from_raster  # noqa
+from .proj4 import (proj4_dict_to_str, proj4_str_to_dict, convert_proj_floats, proj4_radius_parameters)  # noqa
+from .rasterio import get_area_def_from_raster  # noqa
 
 
 def get_area_def(*args, **kwargs):
