+++ pyresample/utils/cartopy.py
@@ -27,7 +27,7 @@ try:
 except ImportError:
     DataArray = np.ndarray
 
-from pyresample.utils._proj4 import proj4_str_to_dict
+from pyresample.utils.proj4 import proj4_str_to_dict
 import cartopy.crs as ccrs
 import shapely.geometry as sgeom
