+++ pyresample/utils/rasterio.py
